+++ src/FrontDoor/FrontDoor/Cmdlets/NewAzureRmFrontDoorMatchConditionObject.cs
@@ -12,16 +12,12 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-using System;
-using System.Collections;
+using System.Linq;
 using System.Management.Automation;
-using System.Net;
 using Microsoft.Azure.Commands.FrontDoor.Common;
 using Microsoft.Azure.Commands.FrontDoor.Models;
-using Microsoft.Azure.Management.FrontDoor;
-using System.Linq;
-using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
+using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 
 namespace Microsoft.Azure.Commands.FrontDoor.Cmdlets
 {
@@ -38,7 +34,8 @@ namespace Microsoft.Azure.Commands.FrontDoor.Cmdlets
         /// 'RequestUri', 'RequestHeader', 'RequestBody'
         /// </summary>
         [Parameter(Mandatory = true, HelpMessage = "Match Variable. Possible values include: 'RemoteAddr', 'RequestMethod', 'QueryString', 'PostArgs','RequestUri', 'RequestHeader', 'RequestBody'")]
-        public PSMatchVariable MatchVariable { get; set; }
+        [PSArgumentCompleter("RemoteAddr", "RequestMethod", "QueryString", "PostArgs", "RequestUri", "RequestHeader", "RequestBody")]
+        public string MatchVariable { get; set; }
 
         /// <summary>
         /// Describes operator to be matched.
@@ -47,13 +44,13 @@ namespace Microsoft.Azure.Commands.FrontDoor.Cmdlets
         /// 'GreaterThanOrEqual', 'BeginsWith', 'EndsWith'
         /// </summary>
         [Parameter(Mandatory = true, HelpMessage = "Describes operator to be matched. Possible values include: 'Any', 'IPMatch', 'GeoMatch', 'Equal', 'Contains', 'LessThan', 'GreaterThan', 'LessThanOrEqual', 'GreaterThanOrEqual', 'BeginsWith', 'EndsWith''")]
-        public PSOperatorProperty OperatorProperty { get; set; }
+        [PSArgumentCompleter("Any", "IPMatch", "GeoMatch", "Equal", "Contains", "LessThan", "GreaterThan", "LessThanOrEqual", "GreaterThanOrEqual", "BeginsWith", "EndsWith")]
+        public string OperatorProperty { get; set; }
 
         /// <summary>
         /// Match value.
         /// </summary>
-        [Parameter(Mandatory = true, HelpMessage = "Match value.")]
-        [ValidateNotNullOrEmpty]
+        [Parameter(Mandatory = false, HelpMessage = "Match value.")]
         public string[] MatchValue { get; set; }
 
         /// <summary>
@@ -67,19 +64,33 @@ namespace Microsoft.Azure.Commands.FrontDoor.Cmdlets
         /// </summary>
         [Parameter(Mandatory = false, HelpMessage = "Describes if this is negate condition or not. Default value is false")]
         public bool NegateCondition { get; set; }
-        
+
         public override void ExecuteCmdlet()
         {
+            ValidateArguments();
+
             var matchCondition = new PSMatchCondition
             {
-               MatchVariable = MatchVariable,
-               MatchValue = MatchValue.ToList(),
-               NegateCondition = !this.IsParameterBound(c => c.NegateCondition)? false : NegateCondition,
-               OperatorProperty = OperatorProperty,
-               Selector = Selector
+                MatchVariable = MatchVariable,
+                MatchValue = MatchValue?.ToList(),
+                NegateCondition = !this.IsParameterBound(c => c.NegateCondition) ? false : NegateCondition,
+                OperatorProperty = OperatorProperty,
+                Selector = Selector
             };
             WriteObject(matchCondition);
         }
-        
+
+        private void ValidateArguments()
+        {
+            if (OperatorProperty == PSOperatorProperty.Any.ToString() && MatchValue != null)
+            {
+                throw new PSArgumentException(nameof(MatchValue));
+            }
+
+            if (OperatorProperty != PSOperatorProperty.Any.ToString() && (MatchValue == null || MatchValue.Length == 0))
+            {
+                throw new PSArgumentNullException(nameof(MatchValue));
+            }
+        }
     }
 }