+++ src/Sql/Sql.Test/ScenarioTests/ManagedInstanceCrudScenarioTests.ps1
@@ -33,6 +33,7 @@ function Test-CreateManagedInstance
  	$vCore = 16
  	$skuName = "GP_Gen4"
 	$collation = "Serbian_Cyrillic_100_CS_AS"
+	$timezoneId = "Central Europe Standard Time"
 	$proxyOverride = "Proxy"
 
  	try
@@ -45,7 +46,7 @@ function Test-CreateManagedInstance
  		$job = New-AzSqlInstance -ResourceGroupName $rg.ResourceGroupName -Name $managedInstanceName `
  			-Location $rg.Location -AdministratorCredential $credentials -SubnetId $subnetId `
   			-LicenseType $licenseType -StorageSizeInGB $storageSizeInGB -Vcore $vCore -SkuName $skuName -Collation $collation `
-			-PublicDataEndpointEnabled -ProxyOverride $proxyOverride -AsJob
+			-TimezoneId $timezoneId -PublicDataEndpointEnabled -ProxyOverride $proxyOverride -AsJob
  		$job | Wait-Job
  		$managedInstance1 = $job.Output
 
@@ -59,6 +60,7 @@ function Test-CreateManagedInstance
 		Assert-AreEqual $managedInstance1.VCores $vCore
 		Assert-AreEqual $managedInstance1.StorageSizeInGB $storageSizeInGB
 		Assert-AreEqual $managedInstance1.Collation $collation
+		Assert-AreEqual $managedInstance1.TimezoneId $timezoneId
 		Assert-AreEqual $managedInstance1.PublicDataEndpointEnabled $true
 		Assert-AreEqual $managedInstance1.ProxyOverride $proxyOverride
  		Assert-StartsWith ($managedInstance1.ManagedInstanceName + ".") $managedInstance1.FullyQualifiedDomainName
@@ -221,11 +223,11 @@ function Test-GetManagedInstance
 		Assert-AreEqual $managedInstance1.VCores $resp1.VCores
 		Assert-AreEqual $managedInstance1.StorageSizeInGB $resp1.StorageSizeInGB
 		
-		$all = Get-AzSqlInstance -ResourceGroupName $rg.ResourceGroupName
+		$all = Get-AzSqlInstance -ResourceGroupName $rg.ResourceGroupName -Name *
 		Assert-AreEqual 1 $all.Count
 
 		# Test getting all managedInstances in all resource groups
-		$all2 = Get-AzSqlInstance
+		$all2 = Get-AzSqlInstance -ResourceGroupName *
 
 		# It is possible that there were existing managedInstances in the subscription when the test was recorded, so make sure
 		# that the managedInstances that we created are retrieved and ignore the other ones.