+++ src/ContainerInstance/ContainerInstance/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 
 ## Upcoming Release
+
+## Version 1.0.1
 * Fixed issue in the -Command parameter of New-AzContainerGroup which added a trailing empty argument
 
 ## Version 1.0.0