+++ src/FrontDoor/FrontDoor/ChangeLog.md
@@ -18,6 +18,15 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+## Version 0.7.4
+* Change enum type parameters to string.
+
+## Version 0.7.3
+* Configure redirect routing rule.
+* Enable/Disable cetificate name check for backend pools
+* Modify WAF policy cmdlets to implement new swagger
+    - Adds new managed rule sets capabilities
+    - Adds redirect action
 
 ## Version 0.7.2
 * Add cmdlets to enable/disable custom domain SSL